web/requirements.txt
flask==2.1.2
plotly==5.7.0
pandas==1.4.2
pyyaml==6.0
pymysql==1.0.2
